Creates DashboardController
